Update example app
