+++ example/lib/src/first_method.dart
@@ -1,0 +1,42 @@
+import 'package:example/src/home.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+
+class MyApp extends StatelessWidget {
+  const MyApp({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    // In first method you only need to wrap [MaterialApp] with [ScreenUtilInit] and that's it
+    return ScreenUtilInit(
+      builder: (context) {
+        return MaterialApp(
+          debugShowCheckedModeBanner: false,
+          // Use this line to prevent extra rebuilds
+          useInheritedMediaQuery: true,
+          title: 'First Method',
+          // You can use the library anywhere in the app even in theme
+          theme: ThemeData(
+            primarySwatch: Colors.blue,
+            textTheme: TextTheme(bodyText2: TextStyle(fontSize: 30.sp)),
+          ),
+          home: HomePage(title: 'First Method'),
+        );
+      },
+    );
+  }
+}
+
+class HomePage extends StatefulWidget {
+  const HomePage({Key? key, required this.title}) : super(key: key);
+
+  final String title;
+
+  @override
+  _HomePageState createState() => _HomePageState();
+}
+
+class _HomePageState extends State<HomePage> {
+  @override
+  Widget build(BuildContext context) => HomePageScaffold(title: widget.title);
+}
